+++ src/Experimental/RAGTools/RAGTools.jl
@@ -16,6 +16,9 @@ using AbstractTrees
 using AbstractTrees: PreOrderDFS
 const PT = PromptingTools
 
+# reexport
+export pprint
+
 ## export trigrams, trigrams_hashed, text_to_trigrams, text_to_trigrams_hashed
 ## export STOPWORDS, tokenize, split_into_code_and_sentences
 include("utils.jl")
@@ -25,21 +28,22 @@ include("api_services.jl")
 
 include("rag_interface.jl")
 
-export ChunkIndex, CandidateChunks
+export ChunkIndex, CandidateChunks, RAGResult
 # export MultiIndex # not ready yet
 include("types.jl")
 
 export build_index, get_chunks, get_embeddings, get_tags
 include("preparation.jl")
 
-export retrieve
+export retrieve, SimpleRetriever, AdvancedRetriever
 export find_closest, find_tags, rerank, rephrase
 include("retrieval.jl")
 
-export airag, build_context, generate!, refine!, answer!, postprocess!
+export airag, build_context!, generate!, refine!, answer!, postprocess!
+export SimpleGenerator, AdvancedGenerator, RAGConfig
 include("generation.jl")
 
-export annotate_support
+export annotate_support, TrigramAnnotater
 include("annotation.jl")
 
 export build_qa_evals, run_qa_evals